consolidate logging in geophires_docs
